compiler/bootstrap: Reverted to using 'uintptr' for bootstrap; Hardcoding FileListing size for cross-compiling to 64 bit

+++ compiler/bootstrap/ecere/bootstrap/File.c
@@ -265,7 +265,7 @@ struct __ecereNameSpace__ecere__com__Method * method;
 } __attribute__ ((gcc_struct));
 } __attribute__ ((gcc_struct));
 
-typedef unsigned int size_t;
+typedef uintptr_t size_t;
 
 void exit(int status);
